feat: create package-lock.json with prod dependencies only
Creates a package-lock.json which only contains the package data for the
dependencies, and not the devDependencies.

devDependencies should anyway not be contained in the lambda layers,
but we want to use the package information from the lockfile, when creating
the lambda layer.

This reduces the size of node_modules significantly:

    npm i --no-audit --ignore-scripts --only=prod --legacy-peer-deps
    du -h
    # 380M    .

With a lockfile that only has the dependencies:

    npm ci --no-audit --ignore-scripts --only=prod --legacy-peer-deps
    du -h
    # 160M    .